interview/src/test/java/uber/_403_Frog_Jump/SolutionITest.java
package uber._403_Frog_Jump;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

/**
 * @author jacka
 * @version 1.0 on 5/16/2021
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int[] stones = {0, 1, 3, 5, 6, 8, 12, 17};
    assertTrue(solution.canCross(stones));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    final int[] stones = {0, 1, 2, 3, 4, 8, 9, 11};
    assertFalse(solution.canCross(stones));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase3(Solution solution) {
    final int[] stones = {0, 1, 3, 5, 6, 8, 12, 17};
    assertTrue(solution.canCross(stones));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}